Add home folder permissions for Flatpak
It's likely that the user wants Kemono Downloader to download into their downloads folder, not the Flatpak's own sandbox filesystem.

Note that the default downloads folder is still in the sandbox, but this commit allows them to select their own downloads folder (and potentially a folder used by a previous instance of Kemono Downloader) without a temporary folder being created.